Use windowDidLoad instead of awakeFromNib.
This is somewhat subjective, but windowDidLoad is a more specific callback (hence better) where awakeFromNib is to be used in lack of better (and generally for members of the nib).

Additionally this removes the call to super which would raise an exception on 10.5 (where NSWindowController does not respond to this method).