Make preferences window resizable.
This is only done when the content view has either a resizable width or height.

The minimum and maximum sizes are hardcoded. I think leaving the maximum size as CGFLOAT_MAX × CGFLOAT_MAX is fine, but minimum size should probably be the initial frame size of the content view (until we require Lion and can leverage the new constraint based layout engine).