Refactoring.
A functional change is that we now commit editing by calling the view controller’s ‘commitEditing’ (rather than change first responder). Additionally we refuse to close window or change selected view controller if committing failed.

As for the refactoring: The selectedViewController property is now read/write and the logic is contained in setSelectedViewController: which will:

 - Commit changes in current selected view controller, failure results in aborting the property change (changing back the selected tool bar item).
 - Setup the new view controller, resize window, change window title, change selected tool bar item, issue proper notifications, etc.