Add initialKeyView protocol method.
This should be implemented if a preferences pane wants to affect which view should be the initial first responder (when switching to the pane).

I took the method signature from the NSPreferencePane class rather than mimic the signature of NSWindow’s initialFirstResponder — I think the former sounds more appropriate, OTOH the latter is likely what most are familiar with, I have no strong preference either way.

I also made it a method rather than a property, mainly because it is optional so we need to use ‘respondsToSelector:’ and I feel that is a semantic mismatch to test for the presence of properties. Though I am inclined to changing it to a property, especially since this is likely something you want to connect in Interface Builder.