Fix the size of Compilation Settings sheet [#250, #201]
The previous default size of the sheet is now too small, and we used
to restore that size for all existing users. We now enforce a certain
reaonable minimum width and height.

Additionally, the sheet used to shink vertically each time it was
opened because we had a title bar enabled for this window in Interface
Builder. What happened was:

- in windowDidLoad we restored the saved size

- when displaying the window as a sheet, Cocoa removed the title bar

- window got smaller by the size of the title bar

- windowDidResize was triggered, and we saved the new smaller size

This commit removes the title bar style to fix that.