windows: Ditch fastJSON, copy the classes that we use
Just because I'm evil! Code copying ftw!

Ah no, just kidding. Because fastJSON parser was choking when fed an
untyped JSON object, and all I really wanted was the underlying
JsonParser class without the surrounding complexity.

I've copied JsonParser and JsonFormatter pretty much intact, and I've
stripped a lot of unnecessary fluff from JsonSerializer, leaving just
the bare-bones serializer of pure Json data types. No surprises if we
ever accidentally try to serialize a C# object.

Seriously, this code is so simple that it's almost silly to have it as
an external dependency.